Issue #4586: Moved resources for JavadocTagContinuationIndentation check test
