+++ src/test/resources/com/puppycrawl/tools/checkstyle/checks/javadoc/javadoctagcontinuationindentation/InputJavadocTagContinuationIndentationGuavaFalsePositive.java
@@ -1,8 +1,8 @@
-package com.puppycrawl.tools.checkstyle.checks.javadoc;
+package com.puppycrawl.tools.checkstyle.checks.javadoc.javadoctagcontinuationindentation;
 
 import java.util.function.Function;
 
-class Foo5 {
+class InputJavadocTagContinuationIndentationGuavaFalsePositive {
 
     /**
      * This class implements the GWT serialization of {@link HashMultimap}.
